# Optimize code to materialize js objects from wasm