--- a/rust/automerge-wasm/Cargo.toml
+++ b/rust/automerge-wasm/Cargo.toml
@@ -34,6 +34,7 @@ serde_bytes = "0.11.5"
 hex = "^0.4.3"
 itertools = "^0.10.3"
 thiserror = "^1.0.16"
+fxhash = "^0.2.1"
 
 [dependencies.wasm-bindgen]
 version = "^0.2.85"