--- a/rust/automerge-wasm/src/value.rs
+++ b/rust/automerge-wasm/src/value.rs
@@ -20,6 +20,10 @@ pub(crate) enum Datatype {
 }
 
 impl Datatype {
+    pub(crate) fn is_seq(&self) -> bool {
+        matches!(self, Self::List | Self::Text)
+    }
+
     pub(crate) fn is_scalar(&self) -> bool {
         !matches!(self, Self::Map | Self::Table | Self::List | Self::Text)
     }