This is primarily code cleanup to replace a recursive algorithm with a non-recursive one.  In the course of writing this I found several low hanging fruit and was able to gleam a 20% performance improvement for the `large-scene` use case with 1M objects.   We also identified an opportunity to get another 25-50% speedup via using a WeakMap to store metadata.

Before

`Load :  8965 ms   Materialize :  6724 ms    -- ./larger-scene.amg`

After

`Load :  7856 ms   Materialize :  5050 ms    -- ./larger-scene.amg`